Added NamedParameterJdbcTemplat: Book, Author

+++ spring-05-books/src/main/java/ru/otus/spring05books/dao/AuthorDao.java
@@ -56,6 +56,6 @@ public interface AuthorDao {
      * Метод getCountOfAuthors получает число авторов, чьи сведения есть в библиотеке
      * @return
      */
-    long getCountOfAuthors();
+    int getCountOfAuthors();
 
 }
