+++ spring-05-books/src/main/java/ru/otus/spring05books/dao/BookDao.java
@@ -61,6 +61,6 @@ public interface BookDao {
      *
      * @return
      */
-    long getCountOfBooks();
+    int getCountOfBooks();
 
 }
