+++ spring-05-books/src/main/java/ru/otus/spring05books/dao/GenreDao.java
@@ -60,6 +60,6 @@ public interface GenreDao {
      *
      * @return
      */
-    long getCountOfGenres();
+    int getCountOfGenres();
 
 }
